Move files into src/ dir
