--- examples/Example/app.json
@@ -1,4 +1,0 @@
-{
-  "name": "Example",
-  "displayName": "Example"
-}
