+++ examples/Example/index.tsx
@@ -1,5 +1,4 @@
 import {AppRegistry} from 'react-native';
-import App from './App';
-import {name as appName} from './app.json';
+import App from './src/App';
 
-AppRegistry.registerComponent(appName, () => App);
+AppRegistry.registerComponent('Example', () => App);
