+++ examples/Example/src/common.tsx
