+++ examples/Example/src/config.tsx
