+++ examples/Example/src/doubleDraggable/index.tsx
