+++ examples/Example/src/doubleScalePinchAndRotate/index.tsx
