+++ examples/Example/src/pagerAndDrawer/index.android.tsx
