+++ examples/Example/src/pagerAndDrawer/index.tsx
