+++ examples/Example/src/panResponder/index.tsx
