+++ examples/Example/src/scaleAndRotate/index.tsx
