+++ examples/Example/src/scaleAndRotate/swmansion.png
