+++ examples/Example/src/swipeable/AppleStyleSwipeableRow.tsx
