+++ examples/Example/src/swipeable/GmailStyleSwipeableRow.tsx
