+++ examples/Example/src/swipeable/index.tsx
